fix(verification): reject redirect-loop API candidates
