+++ public/metagraph/freshness.json
@@ -160,7 +160,7 @@
       "timestamp_field": "candidate_discovery_as_of"
     },
     {
-      "as_of": "2026-06-09T05:36:53.950Z",
+      "as_of": "2026-06-09T06:18:31.409Z",
       "id": "candidate-verification",
       "lane": "candidate-verification",
       "notes": "",
@@ -169,7 +169,7 @@
       "stale_after_hours": 24,
       "stale_behavior": "block",
       "status": "captured",
-      "timestamp": "2026-06-09T05:36:53.950Z",
+      "timestamp": "2026-06-09T06:18:31.409Z",
       "timestamp_field": "verification_as_of"
     },
     {
@@ -228,7 +228,7 @@
     "stale_window_warnings": [
       "surface-health has no observed timestamp; production publish should block."
     ],
-    "verification_as_of": "2026-06-09T05:36:53.950Z",
+    "verification_as_of": "2026-06-09T06:18:31.409Z",
     "verification_generated_at": "1970-01-01T00:00:00.000Z",
     "warning_source_count": 11
   }
