+++ public/metagraph/r2-manifest.json
@@ -1,6 +1,6 @@
 {
   "artifact_count": 23,
-  "artifact_size_bytes": 5363039,
+  "artifact_size_bytes": 5361160,
   "artifacts": [
     {
       "content_type": "application/json",
@@ -16,8 +16,8 @@
       "key": "runs/1970-01-01T00-00-00-000Z/changelog.json",
       "latest_key": "latest/changelog.json",
       "path": "/metagraph/changelog.json",
-      "sha256": "294e5902f3a17ece7bf42a7232be9676561a6470e5ae4bceef0e8ad816097d6b",
-      "size_bytes": 3150,
+      "sha256": "9b9c927259e685489784e34a99eb28f5c131e5354be00e7780e8400703763e3a",
+      "size_bytes": 1349,
       "storage_tier": "dual"
     },
     {
@@ -61,7 +61,7 @@
       "key": "runs/1970-01-01T00-00-00-000Z/freshness.json",
       "latest_key": "latest/freshness.json",
       "path": "/metagraph/freshness.json",
-      "sha256": "6d9886cbf42c50522fc157b9c4558571e5090cbeffae445dd13fecb11156a63c",
+      "sha256": "f3efe62128910a2c9b6e24a67e54ccab48d8731422f2854e94ff80d33eb330a5",
       "size_bytes": 7549,
       "storage_tier": "dual"
     },
@@ -115,7 +115,7 @@
       "key": "runs/1970-01-01T00-00-00-000Z/review/curation.json",
       "latest_key": "latest/review/curation.json",
       "path": "/metagraph/review/curation.json",
-      "sha256": "65f39d7bcbb4019338e4209ea5491cca02974ea29a328c0baf46b677145b697b",
+      "sha256": "32126ce80d66dadf7772d0bf5b53a1d2dabd4cbeb47e21c4965d08ac2ce07ccb",
       "size_bytes": 157066,
       "storage_tier": "dual"
     },
@@ -124,8 +124,8 @@
       "key": "runs/1970-01-01T00-00-00-000Z/review/enrichment-queue.json",
       "latest_key": "latest/review/enrichment-queue.json",
       "path": "/metagraph/review/enrichment-queue.json",
-      "sha256": "8424b4a8a7f82421ac57852b8d5edc16c618e44d6c02d34abf84175ae27573a2",
-      "size_bytes": 366285,
+      "sha256": "3af9494dc39ee00e4d7a3d6d50eed382e4cdd77107f7e2d57da233aef48722ef",
+      "size_bytes": 366207,
       "storage_tier": "dual"
     },
     {
@@ -133,7 +133,7 @@
       "key": "runs/1970-01-01T00-00-00-000Z/review/gap-priorities.json",
       "latest_key": "latest/review/gap-priorities.json",
       "path": "/metagraph/review/gap-priorities.json",
-      "sha256": "d70c604e30686fc807cc4534e390450e171c2606d6228f3179c5f29ccec72edf",
+      "sha256": "279ae4df8bc25ccab0acc8811e3d0c962b044a31a22fae6a9f32c7d1eab4fa8a",
       "size_bytes": 65469,
       "storage_tier": "dual"
     },
@@ -214,7 +214,7 @@
   "bucket_name": "metagraphed-artifacts",
   "contract_version": "2026-06-06.1",
   "full_artifact_count": 1273,
-  "full_artifact_size_bytes": 49158171,
+  "full_artifact_size_bytes": 49159278,
   "full_manifest_key": "latest/r2-manifest.json",
   "full_manifest_run_key": "runs/1970-01-01T00-00-00-000Z/r2-manifest.json",
   "generated_at": "1970-01-01T00:00:00.000Z",
@@ -238,8 +238,8 @@
     "r2": 1250
   },
   "storage_tier_size_bytes": {
-    "dual": 5104079,
+    "dual": 5102200,
     "git": 258960,
-    "r2": 43795132
+    "r2": 43798118
   }
 }
