+++ public/metagraph/review/gap-priorities.json
@@ -957,7 +957,7 @@
       "slug": "sn-114",
       "suggested_next_action": "keep baseline entry and wait for public-source or community intake",
       "surface_count": 11,
-      "verified_candidate_count": 16
+      "verified_candidate_count": 15
     },
     {
       "candidate_count": 6,
