+++ scripts/http-probe-classification.mjs
@@ -1,0 +1,70 @@
+import { isJsonContentType } from "./lib.mjs";
+
+export function classifyHttpProbe(probe, candidate = null) {
+  if (probe.unsafe_url || probe.private_redirect_blocked) {
+    return "unsafe";
+  }
+  if (probe.error === "redirect limit exceeded") {
+    return "unsupported";
+  }
+  if (probe.error_class === "AbortError") {
+    return "timeout";
+  }
+  if (
+    probe.redirect_target &&
+    probe.status_code >= 200 &&
+    probe.status_code < 400
+  ) {
+    if (isContentMismatch(probe, candidate)) {
+      return "content-mismatch";
+    }
+    return "redirected";
+  }
+  if (probe.status_code >= 200 && probe.status_code < 400) {
+    if (isContentMismatch(probe, candidate)) {
+      return "content-mismatch";
+    }
+    return "live";
+  }
+  if (probe.status_code === 429) {
+    return "rate-limited";
+  }
+  if ([401, 403].includes(probe.status_code)) {
+    return "auth-required";
+  }
+  if ([404, 410].includes(probe.status_code)) {
+    return "dead";
+  }
+  if (probe.status_code >= 500) {
+    return "transient";
+  }
+  return "unsupported";
+}
+
+export function isContentMismatch(probe, candidate) {
+  if (!candidate || !probe.ok) {
+    return false;
+  }
+  if (candidate.kind === "openapi") {
+    return !isJsonContentType(probe.content_type);
+  }
+  if (candidate.kind === "subnet-api") {
+    return !isMachineReadableApiContentType(probe.content_type);
+  }
+  if (candidate.kind === "sse") {
+    return !String(probe.content_type || "")
+      .toLowerCase()
+      .includes("text/event-stream");
+  }
+  return false;
+}
+
+function isMachineReadableApiContentType(contentType) {
+  const normalized = String(contentType || "").toLowerCase();
+  return (
+    normalized.includes("json") ||
+    normalized.includes("text/plain") ||
+    normalized.includes("text/event-stream") ||
+    normalized.includes("application/octet-stream")
+  );
+}
