+++ scripts/lib.mjs
@@ -17,6 +17,24 @@ export const r2StagingRoot = path.join(repoRoot, R2_STAGING_RELATIVE_ROOT);
 export const generatedSourceRoot = path.join(repoRoot, "dist/metagraph-source");
 
 const credentialedUrlParams = new Set([
+  "app_domain",
+  "authuser",
+  "client_id",
+  "code_challenge",
+  "code_challenge_method",
+  "continue",
+  "dsh",
+  "flowname",
+  "nonce",
+  "opparams",
+  "part",
+  "prompt",
+  "rart",
+  "redirect_uri",
+  "response_type",
+  "scope",
+  "service",
+  "state",
   "x-amz-credential",
   "x-amz-signature",
   "x-amz-security-token",
