+++ scripts/verify-candidates.mjs
@@ -12,6 +12,10 @@ import {
   stableStringify,
   writeJson,
 } from "./lib.mjs";
+import {
+  classifyHttpProbe,
+  isContentMismatch,
+} from "./http-probe-classification.mjs";
 import { preservePreviousGithubMetadata } from "./verification-quality.mjs";
 
 const args = new Set(process.argv.slice(2));
@@ -389,6 +393,19 @@ async function probeUrl(url, method, accept, redirectCount = 0) {
       };
     }
 
+    if ([301, 302, 303, 307, 308].includes(response.status) && location) {
+      const redirectTarget = new URL(location, url).toString();
+      await response.body?.cancel();
+      return {
+        ok: false,
+        error: "redirect limit exceeded",
+        latency_ms: latencyMs,
+        method_tested: method,
+        redirect_target: redactCredentialedUrl(redirectTarget),
+        status_code: response.status,
+      };
+    }
+
     await response.body?.cancel();
     return {
       ok: response.ok,
@@ -439,74 +456,8 @@ async function fetchJson(url, headers = {}) {
   }
 }
 
-function classifyHttpProbe(probe, candidate = null) {
-  if (probe.unsafe_url || probe.private_redirect_blocked) {
-    return "unsafe";
-  }
-  if (probe.error_class === "AbortError") {
-    return "timeout";
-  }
-  if (
-    probe.redirect_target &&
-    probe.status_code >= 200 &&
-    probe.status_code < 400
-  ) {
-    if (isContentMismatch(probe, candidate)) {
-      return "content-mismatch";
-    }
-    return "redirected";
-  }
-  if (probe.status_code >= 200 && probe.status_code < 400) {
-    if (isContentMismatch(probe, candidate)) {
-      return "content-mismatch";
-    }
-    return "live";
-  }
-  if (probe.status_code === 429) {
-    return "rate-limited";
-  }
-  if ([401, 403].includes(probe.status_code)) {
-    return "auth-required";
-  }
-  if ([404, 410].includes(probe.status_code)) {
-    return "dead";
-  }
-  if (probe.status_code >= 500) {
-    return "transient";
-  }
-  return "unsupported";
-}
-
 function isPromotable(result) {
   return ["live", "redirected"].includes(result.classification);
-}
-
-function isContentMismatch(probe, candidate) {
-  if (!candidate || !probe.ok) {
-    return false;
-  }
-  if (candidate.kind === "openapi") {
-    return !isJsonContentType(probe.content_type);
-  }
-  if (candidate.kind === "subnet-api") {
-    return !isMachineReadableApiContentType(probe.content_type);
-  }
-  if (candidate.kind === "sse") {
-    return !String(probe.content_type || "")
-      .toLowerCase()
-      .includes("text/event-stream");
-  }
-  return false;
-}
-
-function isMachineReadableApiContentType(contentType) {
-  const normalized = String(contentType || "").toLowerCase();
-  return (
-    normalized.includes("json") ||
-    normalized.includes("text/plain") ||
-    normalized.includes("text/event-stream") ||
-    normalized.includes("application/octet-stream")
-  );
 }
 
 function scoreCandidate(candidate, probe) {
